+++ apps/backend/backend_main.c
@@ -497,6 +497,12 @@ main(int    argc,
     if (help)
 	usage(h, argv[0]);
 
+#ifndef HAVE_LIBXML2
+    if (strcmp(clicon_yang_regexp(h), "libxml2")==0){
+	clicon_err(OE_FATAL, 0, "CLICON_YANG_REGEXP set to libxml2, but HAVE_LIBXM2 not set (Either change CLICON_YANG_REGEXP to posix, or install libxml2?))");
+	goto done;
+    }
+#endif
     /* Check pid-file, if zap kil the old daemon, else return here */
     if ((pidfile = clicon_backend_pidfile(h)) == NULL){
 	clicon_err(OE_FATAL, 0, "pidfile not set");