+++ lib/clixon/clixon_yang_type.h
@@ -67,7 +67,7 @@ int        ys_resolve_type(yang_stmt *ys, void *arg);
 int        yang2cv_type(char *ytype, enum cv_type *cv_type);
 char      *cv2yang_type(enum cv_type cv_type);
 yang_stmt *yang_find_identity(yang_stmt *ys, char *identity);
-int        ys_cv_validate(cg_var *cv, yang_stmt *ys, char **reason);
+int        ys_cv_validate(clicon_handle h, cg_var *cv, yang_stmt *ys, char **reason);
 int        clicon_type2cv(char *type, char *rtype, yang_stmt *ys, enum cv_type *cvtype);
 int        yang_type_get(yang_stmt *ys, char **otype, yang_stmt **restype, 
 			 int *options, cvec **cvv, char **pattern,